Switch to ES6 syntax
