--- assets/src/scripts/theme.js
@@ -1,7 +1,0 @@
-( function( document ) {
-
-	var pageHistory = require("./inc/history.js");
-
-	pageHistory.init();
-
-} )( document );
