+++ gulpfile.js
@@ -61,7 +61,7 @@ gulp.task( 'lint', function () {
 // Watch for changes in JS/CSS.
 gulp.task('watch', function() {
 	gulp.watch( 'assets/src/styles/**/*.scss', ['styles'] );
-	gulp.watch( ['assets/src/scripts/**/*.js'], ['js'] );
+	gulp.watch( [ 'assets/src/scripts/**/*.js', 'assets/src/scripts/**/*.jsx' ], ['js'] );
 });
 
 // Tasks
