+++ package.json
@@ -10,7 +10,9 @@
     "autoprefixer": "^6.3.4",
     "babel-core": "^6.10.4",
     "babel-loader": "^6.2.4",
+    "babel-preset-es2015": "^6.9.0",
     "babel-preset-react": "^6.11.1",
+    "diff-match-patch": "^1.0.0",
     "exports-loader": "^0.6.3",
     "gulp": "^3.9.0",
     "gulp-postcss": "^6.1.0",
